[BUILD] Removed AC_CHECK_FUNCS for asprintf
This caused an issue with android-mips builds. The function exists (so
the check should work !), but the check fails on link error for the
libarsal.so and libarstream.so.

The linker does not find the '__android_log_vprint' function linked
from libarsal.so, even if libarsal.so has a valid dependency to
liblog.so. Same for the 'sqrt' function in libarstream.so, with a
proper dep to libm.so

Change-Id: I634fa3f8cb147fca4cce5f44f8e941d4b600b3b5
Reviewed-on: https://gerrit.parrot.biz/25049
Reviewed-by: Djavan Bertrand <djavan.bertrand@parrot.com>