[DEV] Use a default port first then try with a random one
Instead of using a random port to open a socket for
the messages or for the stream and its control, first
try to use a default port. If this port is already used,
try a random one.

This fixes the bug that made the quick reconnection (< 5sec)
impossible because the drone is not reading the json
(containing the ports). However, it does not fix this bug if a
random port was selected.

Change-Id: I9616534a73fa7a994d2d7f6fe878e8641e79f028
Reviewed-on: https://gerrit.parrot.biz/69883
Reviewed-by: Maxime Maitre <maxime.maitre@parrot.com>